+++ src/rofi_rbw/typer/xdotool.py
@@ -1,7 +1,7 @@
 from subprocess import run
 
 from ..abstractionhelper import is_installed, is_wayland
-from .typer import Typer, Key
+from .typer import Key, Typer
 
 
 class XDoToolTyper(Typer):